requirements.txt
Django==1.10.5
django-import-export==0.5.1
djangorestframework==3.5.3
docutils==0.12
graphviz==0.5.2
httplib2==0.9.2
idna==2.1
incremental==16.10.1
ipython-genutils==0.1.0
lxml==3.6.4
matplotlib==1.5.3
mysqlclient==1.3.8
oauth2==1.9.0.post1
oauthlib==2.0.0
pickleshare==0.7.4
Pillow==3.4.2
pycparser==2.17
pycurl==7.43.0
PyDispatcher==2.0.5
pydot==1.2.3
pygame==1.9.2b1
Pygments==2.1.3
pyparsing==2.1.10
pypiwin32==219
python-dateutil==2.5.3
pytz==2016.7
queuelib==1.4.2
requests==2.12.1
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
suit==2.0.2

